feat: implement image generation handlers and request types for FMLX and Stable Diffusion
